requirements.txt
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
docx2txt>=0.8
pypdf>=3.0.0
openpyxl>=3.1.0
pandas>=2.0.0
streamlit>=1.32.0
